update api types
