+++ packages/lodestar-types/src/types/api.ts
@@ -1,15 +1,35 @@
 /* eslint-disable @typescript-eslint/interface-name-prefix */
-import {BLSPubkey, BLSSignature, CommitteeIndex, Slot, Uint64} from "./primitive";
+import {BLSPubkey, BLSSignature, CommitteeIndex, Number64, Slot, Uint64} from "./primitive";
 import {Fork} from "./misc";
 
 export interface SubscribeToCommitteeSubnetPayload {
   slot: Slot;
   slotSignature: BLSSignature;
-  committeeIndex: CommitteeIndex;
+  attestationCommitteeIndex: CommitteeIndex;
   aggregatorPubkey: BLSPubkey;
 }
 
 export interface ForkResponse {
   chainId: Uint64;
   fork: Fork;
+}
+
+export interface AttesterDuty {
+  // The validator's public key, uniquely identifying them
+  validatorPubkey: BLSPubkey;
+  // used to determine if validator is aggregator
+  aggregatorModulo: Number64;
+  // The slot at which the validator must attest
+  attestationSlot: Slot;
+
+  committeeIndex: CommitteeIndex;
+}
+
+export interface SyncingStatus {
+  // The block at which syncing started (will only be reset, after the sync reached his head)
+  startingBlock: Uint64;
+  // Current Block
+  currentBlock: Uint64;
+  // The estimated highest block, or current target block number
+  highestBlock: Uint64;
 }
